+++ test/integration/polyfill-node-modules/pages/node-browser-polyfills.js
